Make `TestError` an effectively-never instantiated `struct`
One of clients of this library return errors from their test functions
and explicitly `into()` them. Rust nightly started issuing lint failures
for `TestError` being uninhabited (and as such the code being unreachable):

```
note: this expression has type `testresult::TestError`, which is uninhabited
   --> nethsm/src/user.rs:804:31
    |
804 |             Err(error) => Err(format!("Expected to fail with a Error::InvalidUserIds but got a different error instead:\n{error}").into()),
    |                               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

warning: unreachable expression
   --> nethsm/src/user.rs:805:27
    |
805 |             Ok(user_id) =>Err(format!("Expected to fail with a Error::InvalidUserIds but succeeded instead:\n{user_id}").into())
    |                           ^^^^-------------------------------------------------------------------------------------------------^
    |                           |   |
    |                           |   any code following this expression is unreachable
    |                           unreachable expression
```

This PR converts the `TestError` to a struct that is never instantiated.
Additionally it marks it as `#[non_exhaustive]` so that it is not possible
to instantiate an object of this type outside of our crate.

Signed-off-by: Wiktor Kwapisiewicz <wiktor@metacode.biz>